:sparkles: [EmittenCommon] Easier to extend common use-case
