+++ src/index.ts
@@ -1,4 +1,5 @@
 export {Emitten} from './Emitten';
+export {EmittenCommon} from './EmittenCommon';
 export {EmittenProtected} from './EmittenProtected';
 
 export type {
